+++ Modules/DataModel/Sources/PocketCastsDataModel/Private/DB/FMDB/FMDBQueue.swift
@@ -1,0 +1,32 @@
+import Foundation
+import FMDB
+
+final class FMDBQueue: PCDBQueue {
+    private let fmdbQueue: FMDatabaseQueue
+
+    init(fmdbQueue: FMDatabaseQueue) {
+        self.fmdbQueue = fmdbQueue
+    }
+
+    func inDatabase(_ block: (any PCDatabase) -> Void) {
+        withoutActuallyEscaping(block) { block in
+            fmdbQueue.inDatabase { db in
+                let dbWrapper = FMDBDatabase(fmdbDatabase: db)
+                block(dbWrapper)
+            }
+        }
+    }
+
+    func inTransaction(_ block: (any PCDatabase, UnsafeMutablePointer<ObjCBool>) -> Void) {
+        withoutActuallyEscaping(block) { block in
+            fmdbQueue.inTransaction { db, rollback in
+                let dbWrapper = FMDBDatabase(fmdbDatabase: db)
+                block(dbWrapper, rollback)
+            }
+        }
+    }
+
+    func close() {
+        fmdbQueue.close()
+    }
+}