+++ Modules/DataModel/Sources/PocketCastsDataModel/Private/Managers/EndOfYearDataManager.swift
@@ -20,7 +20,7 @@ class EndOfYearDataManager {
     /// If the user is eligible to see End of Year stats
     ///
     /// All it's needed is a single episode listened for more than 5 minutes.
-    func isEligible(in year: Int, dbQueue: FMDatabaseQueue) -> Bool {
+    func isEligible(in year: Int, dbQueue: PCDBQueue) -> Bool {
         var isEligible = false
 
         dbQueue.inDatabase { db in
@@ -55,7 +55,7 @@ class EndOfYearDataManager {
     /// If this is not true, we check for the total number of items of
     /// this year. If the number is less than or equal 100, we assume they
     /// have the full history.
-    func isFullListeningHistory(in year: Int, dbQueue: FMDatabaseQueue) -> Bool {
+    func isFullListeningHistory(in year: Int, dbQueue: PCDBQueue) -> Bool {
         var isFullListeningHistory = false
 
         dbQueue.inDatabase { db in
@@ -81,7 +81,7 @@ class EndOfYearDataManager {
     }
 
     /// Returns the number of episodes we have for this year
-    func numberOfEpisodes(year: Int, dbQueue: FMDatabaseQueue) -> Int {
+    func numberOfEpisodes(year: Int, dbQueue: PCDBQueue) -> Int {
         var numberOfEpisodes: Int = 0
 
         dbQueue.inDatabase { db in
@@ -110,7 +110,7 @@ class EndOfYearDataManager {
     }
 
     /// Returns the approximate listening time for the current year
-    func listeningTime(in year: Int, dbQueue: FMDatabaseQueue) -> Double? {
+    func listeningTime(in year: Int, dbQueue: PCDBQueue) -> Double? {
         var listeningTime: Double?
 
         dbQueue.inDatabase { db in
@@ -133,7 +133,7 @@ class EndOfYearDataManager {
     /// Returns all the categories the user has listened to podcasts
     ///
     /// The returned array is ordered from the most listened to the least
-    func listenedCategories(in year: Int, dbQueue: FMDatabaseQueue) -> [ListenedCategory] {
+    func listenedCategories(in year: Int, dbQueue: PCDBQueue) -> [ListenedCategory] {
         var listenedCategories: [ListenedCategory] = []
 
         dbQueue.inDatabase { db in
@@ -180,7 +180,7 @@ class EndOfYearDataManager {
 
     /// Return the number of podcasts and episodes listened
     ///
-    func listenedNumbers(in year: Int, dbQueue: FMDatabaseQueue) -> ListenedNumbers {
+    func listenedNumbers(in year: Int, dbQueue: PCDBQueue) -> ListenedNumbers {
         var listenedNumbers = ListenedNumbers(numberOfPodcasts: 0, numberOfEpisodes: 0)
 
         dbQueue.inDatabase { db in
@@ -210,7 +210,7 @@ class EndOfYearDataManager {
     }
 
     /// Return the top podcasts ordered by number of played episodes
-    func topPodcasts(in year: Int, dbQueue: FMDatabaseQueue, limit: Int = 5) -> [TopPodcast] {
+    func topPodcasts(in year: Int, dbQueue: PCDBQueue, limit: Int = 5) -> [TopPodcast] {
         var allPodcasts = [TopPodcast]()
         dbQueue.inDatabase { db in
             do {
@@ -249,7 +249,7 @@ class EndOfYearDataManager {
     }
 
     /// Return the longest listened episode
-    func longestEpisode(in year: Int, dbQueue: FMDatabaseQueue) -> Episode? {
+    func longestEpisode(in year: Int, dbQueue: PCDBQueue) -> Episode? {
         var episode: Episode?
         dbQueue.inDatabase { db in
             do {
@@ -275,7 +275,7 @@ class EndOfYearDataManager {
     }
 
     /// Given a list of UUIDs, return which UUIDs are present on the database
-    func episodesThatExist(year: Int, dbQueue: FMDatabaseQueue, uuids: [String]) -> [String] {
+    func episodesThatExist(year: Int, dbQueue: PCDBQueue, uuids: [String]) -> [String] {
         var episodes: [String] = []
 
         dbQueue.inDatabase { db in
@@ -300,7 +300,7 @@ class EndOfYearDataManager {
         return episodes
     }
 
-    private func numberOfItemsInListeningHistory(in year: Int, db: FMDatabase) -> Int {
+    private func numberOfItemsInListeningHistory(in year: Int, db: PCDatabase) -> Int {
         var numberOfItemsInListeningHistory = 0
 
         do {
@@ -325,7 +325,7 @@ class EndOfYearDataManager {
     }
 
     /// Returns the approximate listening time for the current year
-    func yearOverYearListeningTime(in year: Int, dbQueue: FMDatabaseQueue) -> YearOverYearListeningTime {
+    func yearOverYearListeningTime(in year: Int, dbQueue: PCDBQueue) -> YearOverYearListeningTime {
         var listeningTimeThisYear: Double = 0
         var listeningTimePreviousYear: Double = 0
 
@@ -352,7 +352,7 @@ class EndOfYearDataManager {
 
     /// Returns the number of episodes started and finished
     /// The episode is considered completed if it was played at least 90%
-    func episodesStartedAndCompleted(in year: Int, dbQueue: FMDatabaseQueue) -> EpisodesStartedAndCompleted {
+    func episodesStartedAndCompleted(in year: Int, dbQueue: PCDBQueue) -> EpisodesStartedAndCompleted {
         var started: Int = 0
         var completed: Int = 0