+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 7.85
 -----
-
+- Revert episode detail screen dismiss action when archiving [#2802](https://github.com/Automattic/pocket-casts-ios/pull/2802)
 
 7.84
 -----
@@ -17,7 +17,6 @@
 7.82.1
 -----
 - Fix analytics bug [#2769](https://github.com/Automattic/pocket-casts-ios/pull/2769)
-
 
 7.82
 -----